pyproject.toml
[tool.poetry]
name = "atocoder-review-bot"
version = "0.1.0"
description = ""
authors = ["shunsei"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-dotenv = "^1.0.0"
line-bot-sdk = "^3.7.0"
notion-client = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
